fix(pkg/listen): make the code for /api/analysis expect a manifest implementing listentype.AnalysisRequester
